requirements.txt
flask
yt-dlp
openai-whisper
gTTS
requests
ffmpeg-python
torch
